+++ src/components/bookTagSearch.js
@@ -1,13 +1,14 @@
 import React from "react";
 import { useState, useEffect } from "react";
-import { useNavigate, useParams } from "react-router-dom";
+import { useParams } from "react-router-dom";
 import { getBooksByTag } from "../api/api";
+import { ListBooks } from "./exports"
 
 const BookTagSearch = () => {
   const [listBooks, setListBooks] = useState([]);
-  const [itemIndex, setItemIndex] = useState(null);
+
   const { tagName } = useParams();
-  const navigate = useNavigate();
+
 
   useEffect(() => {
     Promise.all([getBooksByTag(tagName)]).then(([bookResults]) => {
@@ -20,40 +21,10 @@ const BookTagSearch = () => {
     });
   }, []);
 
-  function buttonHandler(e, nav) {
-    e.preventDefault();
-    navigate(nav);
-  }
-
-
   return (
     <div>
-        <h2>{`Books marked as ${tagName}`}</h2>
-      {listBooks.map((book, i) => {
-        return (
-          <div key={book.id} onMouseEnter={() => setItemIndex(i)}
-          onMouseLeave={() => setItemIndex(null)}>
-            <div>
-              <img src={book.bookImage} />
-            </div>
-            <div>
-              <b>Title:</b> {book.title}
-            </div>
-            {itemIndex === i && (
-                <div>
-                  <button
-                    onClick={(e) => {
-                      buttonHandler(e, `/books/${book.id}`);
-                    }}
-                  >
-                    Details
-                  </button>
-                  <button>Add to Cart</button>
-                </div>
-              )}
-          </div>
-        );
-      })}
+        
+    <ListBooks allBooks={listBooks} />
     </div>
   );
 };
