+++ src/components/books.js
@@ -1,12 +1,12 @@
 import React from "react";
 import { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import { getAllBooks, addToCart } from "../api/api";
+import { getAllBooks } from "../api/api";
+import { ListBooks } from "./exports"
 
 const Books = () => {
   const [allBooks, setAllBooks] = useState([]);
-  const [itemIndex, setItemIndex] = useState(null);
-  const navigate = useNavigate();
+
 
   useEffect(() => {
     Promise.all([getAllBooks()]).then(([allBooksResults]) => {
@@ -18,58 +18,7 @@ const Books = () => {
     });
   }, []);
 
-  function buttonHandler(e, nav) {
-    e.preventDefault();
-    navigate(nav);
-  }
-
-  return (
-    <div className="container">
-      <h1>SHOP ALL BOOKS</h1>
-      <div className="booksList">
-        {allBooks.map((book, i) => {
-          return (
-            <div key={book.id} className="booksMap">
-              <div
-                className="booksImage"
-                onMouseEnter={() => setItemIndex(i)}
-                onMouseLeave={() => setItemIndex(null)}
-                onClick={(e) => {
-                  buttonHandler(e, `/books/${book.id}`);
-                }}
-              >
-                <img src={book.bookImage} />
-              </div>
-              <div className="booksWords">
-                <div>
-                  <div className="booksTitle">
-                    <b>{book.title}</b>
-                  </div>
-                  <div className="booksAuthor">
-                    By: {book.authorFirstName} {book.authorLastName}
-                  </div>
-                </div>
-                <div className="booksPrice">${book.price}</div>
-                <div className="booksButtons">
-                  <button
-                    onClick={async (e) => {
-                      e.preventDefault();
-                      book.quantity = 1
-                      console.log(book)
-                      const add = await addToCart(book);
-                      alert(add.message)
-                    }}
-                  >
-                    ADD TO CART
-                  </button>
-                </div>
-              </div>
-            </div>
-          );
-        })}
-      </div>
-    </div>
-  );
-};
+  return( <ListBooks allBooks={allBooks}/>)
+  };
 
 export default Books;
