+++ src/components/exports.js
@@ -1,11 +1,12 @@
 export { default as Home } from "./home";
 export { default as NavBar } from "./navBar";
 export { default as Books } from "./books";
+export { default as ListBooks } from "./listBooks"
 export { default as BookDetails } from "./bookDetails";
 export { default as BookTagSearch } from "./bookTagSearch";
-export {default as Register} from "./register";
-export {default as Cart } from "./cart";
-export {default as Login} from "./login";
-export {default as Profile} from   "./profile"
-export { default as CartItem } from "./cartItem"
+export { default as Register } from "./register";
+export { default as Cart } from "./cart";
+export { default as Login } from "./login";
+export { default as Profile } from "./profile";
+export { default as CartItem } from "./cartItem";
 // export {default as Homepage} from "./homepage"
