+++ src/index.js
@@ -19,7 +19,6 @@ const [loggedIn, setLoggedIn] = useState(localStorage.getItem("token"))
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/books" element={<Books />} />
-        <Route path="/books" element={<Books />} />
         <Route path="/books/:bookId" element={<BookDetails />} />
         <Route path="/booktag/:tagName" element={<BookTagSearch />} />
         <Route path="/cart" element={<Cart />} />
